Updated my gql strings to correct files and folders.
