+++ client/src/components/Messages/index.js
@@ -1,77 +1,26 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
 import { storeDog, getDog } from "../../slices/dogSlice";
-import { useMutation, useSubscription, gql, useQuery } from '@apollo/client';
+import { useMutation, useSubscription, useQuery } from '@apollo/client';
+// Material UI Components
 import Container from '@mui/material/Container';
 import Chip from '@mui/material/Chip';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
+// GraphQL Strings
+import { GET_CONVERSATION_BY_ID } from '../../utils/queries';
+import { POST_MESSAGE } from '../../utils/mutations';
+import { GET_MESSAGES_SUB } from '../../utils/subscriptions';
 
 
-
-
-const GET_MESSAGES_SUB = gql`
-subscription MessageSent($conversationId: ID!) {
-  messageSent(conversationId: $conversationId) {
-    _id
-    dogIds
-    messages {
-      messageId
-      message
-      createdAt
-      dogId
-    }
-  }
-}
-`;
-
-const POST_MESSAGE = gql`
-mutation Mutation($conversationId: ID!, $message: PostMessage) {
-  newMessage(conversationId: $conversationId, message: $message) {
-    _id
-    dogIds
-    messages {
-      messageId
-      dogId
-      message
-      createdAt
-    }
-  }
-}
-`;
-
-const GET_CONVERSATION_BY_ID = gql`
-query GetConversationById($conversationId: ID!) {
-  getConversationById(conversationId: $conversationId) {
-    messageId
-    dogId
-    message
-    createdAt
-  }
-}
-`;
-
-const Text = ({ conversationId }) => {
+const Text = ({ messages }) => {
   const dispatch = useDispatch();
   const dog = useSelector(getDog);
 
-  const convoQuery = useQuery(GET_CONVERSATION_BY_ID,
-    { variables: { conversationId } });
-
-  let convo = convoQuery.data?.getConversationById || [];
-
-  const { data } = useSubscription(GET_MESSAGES_SUB,
-    { variables: { conversationId } },
-  );
-
-  console.log(data, 'DATAAAAAAAAAAAA!!');
-  if (!data) {
-    return null;
-  }
   return (
     <div style={{ marginBottom: "5rem" }}>
-      {convo.map(({ dogId, message }) => {
+      {messages.map(({ dogId, message }) => {
         return (
           <div key={dogId} style={{ textAlign: dog._id === dogId ? "right" : "left" }}>
             <p style={{ marginBottom: "0.3rem" }}>{dogId}</p>
@@ -83,11 +32,28 @@ const Text = ({ conversationId }) => {
   )
 }
 
-const Messages = () => {
+const Messages = ({ conversationId }) => {
+  // Local State
   const [user, setUser] = useState("Victoria");
   const [text, setText] = useState("");
-  const [messages, setMessages] = useState("")
-  const [postMessage] = useMutation(POST_MESSAGE)
+
+  // GraphQL Hooks
+  const [postMessage] = useMutation(POST_MESSAGE);
+
+  const convoQuery = useQuery(GET_CONVERSATION_BY_ID,
+    { variables: { conversationId } });
+
+  let messages = convoQuery.data?.getConversationById || [];
+
+  const { data } = useSubscription(GET_MESSAGES_SUB,
+    { variables: { conversationId } },
+  );
+
+  console.log(data, 'DATAAAAAAAAAAAA!!');
+  if (!data) {
+    return null;
+  }
+  
   const sendMessage = async () => {
     const PostMessage = {
       dogId: "633803594950ea4a2c76c2b6",
@@ -101,7 +67,6 @@ const Messages = () => {
         }
       });
       console.log(data);
-      setMessages(data);
       setText("");
     } else {
       alert("Missing fields!")
@@ -112,7 +77,7 @@ const Messages = () => {
   return (
     <Container>
       <h3>Welcome to DevThoughts! A simple chat app for the GraphQL series!</h3>
-      <Text conversationId={'6337e95a5223045e30bf0203'} />
+      <Text messages={messages} />
       <Grid container spacing={2}>
         <Grid item xs={3}>
           <TextField onChange={(e) => {
