+++ client/src/utils/mutations/index.js
@@ -1,11 +1,13 @@
 //import mutations here
 import { UPLOAD_DOG_IMAGES, UPLOAD_OWNER_IMAGES } from "./imageMutations";
 import { SIGNUP_USER } from "./ownerMutations";
+import { POST_MESSAGE } from "./conversationMutations";
 
 
 //export mutations here, add to the list
 export {
    UPLOAD_DOG_IMAGES,
    UPLOAD_OWNER_IMAGES,
-   SIGNUP_USER
+   SIGNUP_USER,
+   POST_MESSAGE,
 }
