+++ client/src/utils/queries/index.js
@@ -1,11 +1,14 @@
 //import queries here
 //import { QUERY1, QUERY2 } from "./someQueryFile";
 import { GET_ALL_DOGS_BY_OWNER_ID } from "./dogQueries";
-import { GET_CONVERSATIONS_BY_DOG_ID } from './conversationQueries';
+import {
+   GET_CONVERSATIONS_BY_DOG_ID,
+   GET_CONVERSATION_BY_ID } from './conversationQueries';
 
 export {
    GET_ALL_DOGS_BY_OWNER_ID,
-   GET_CONVERSATIONS_BY_DOG_ID
+   GET_CONVERSATIONS_BY_DOG_ID,
+   GET_CONVERSATION_BY_ID
 }
 
 //export queries here, add to the list
