+++ client/src/utils/subscriptions/index.js
@@ -1,0 +1,5 @@
+import { GET_MESSAGES_SUB } from "./conversationSubscriptions";
+
+export {
+    GET_MESSAGES_SUB,
+}
